+++ ouroboros-network-framework/CHANGELOG.md
@@ -7,6 +7,9 @@
 * Addapted to `network-mux` changes in https://github.com/IntersectMBO/ouroboros-network/pull/4999
 * Addapted to `network-mux` changes in https://github.com/IntersectMBO/ouroboros-network/pull/4997
 * Removed deprecated `Ouroboros.Network.Channel.{to,from}Channel` functions.
+* Renamed `requestOutboundConnection` to `acquireOutboundConnection` and
+  `unregister{Inbound,Outbound}Connection` to `release{Inbound,Outbound}Connection`.
+  `AssertionLocation` constructors were renamed as well.
 
 ### Non-breaking changes